Make JSON escape policy more configurable.
Refactor the JSON string escaping policy to be more configurable. Right
now we only have the minimal escaping option built in, but we can easily
implement more whenever we want.

Also I think there might have been a bug in the `HaveEscaped` case where
we didn't transition to the back to the `NotEscaping` state. That's
fixed now. We never hit it anyways.